cursorquit/requirements.txt
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
livekit-plugins-deepgram>=0.6.17,<1.0.0
livekit-plugins-elevenlabs>=0.7.0,<1.0.0
python-dotenv~=1.0
